requirements.txt
backcall==0.1.0
chardet==3.0.4
decorator==4.3.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
parso==0.3.1
pdfminer.six==20181108
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycryptodome==3.7.2
Pygments==2.3.1
six==1.12.0
sortedcontainers==2.1.0
traitlets==4.3.2
wcwidth==0.1.7
